Major changes to Track.c, minor to CreateWeakPoints.c and others
  1) Elastic eqm flagging now coordinated between parallel processes
  2) Threshold used to detect eqm now given as absolute (dbl) value in input.xml

Details:

1a)  Flagging in Track.c used to detect slowing down of surface motions without regard for parallel processes.  This led to termination of processing in one thread when eqm was detected in that thread, while other threads kept going.

Now we use a "consensus" global (between threads) elastic eqm flag that is a logical AND of all the individual thread flags.  This flag can be set in the input xml to ensure that we don't wait for elastic eqm and weak points are created immediately.

1b)  The local time step limit is set to the current time step when the consensus flag is TRUE.  This is another change: before, simulation was set to continue by another "dump" time steps to ensure printing of the final state (which became annoying when plastic strain set in quickly).

NB: the parallelization of tracking has not been tested yet.  This revsion is therefore a "test" commit.

2)  The thresholds used to start tracking and identify elastic eqm are now given explicitly (absolute not relative values) in the input xml.  Previously the initial velocity, acceln values were used to normalize their instantaneous values and thresholding was done against these relative values.

Thus, all previous input xml files will not work correctly with the new version.  Changes must be made to startThreshold and stopThreshold variables.  Defaults are 1e-2 and 1e-3.