Follow-up to changes to tracking code to make it work in parallel.
Bug fix: was doing an MPI_Allreduce in a conditional block that would not be reached in all threads, which of course is a big no-no - the run eventually hangs and then bails (on Ranger) as the threads become terminally unsyncable.

Corrected by placing MPI_Allreduce call - which checks to see if all threads are elastically stabilized (equilibrated) or if they can't comment - in a block that will be reached in all threads.